+++ src/main/java/cloud/chlora/pipeline/notification/domain/model/BatteryNotificationThreshold.java
@@ -12,20 +12,20 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public enum BatteryNotificationThreshold {
 
-    LOW(20f, NotificationSeverity.WARNING),
-    VERY_LOW(10f, NotificationSeverity.WARNING),
-    CRITICAL(5f, NotificationSeverity.CRITICAL);
+    LOW(20, NotificationSeverity.WARNING),
+    VERY_LOW(10, NotificationSeverity.WARNING),
+    CRITICAL(5, NotificationSeverity.CRITICAL);
 
-    private final float percent;
+    private final int percent;
     private final NotificationSeverity severity;
 
     /**
      * Menentukan threshold tertinggi yang cocok untuk battery level tertentu.
-     * Contoh: batteryLevel 8.5 → VERY_LOW (karena < 10, tapi belum < 5).
-     *         batteryLevel 3.0 → CRITICAL
-     *         batteryLevel 25  → empty (tidak perlu notifikasi)
+     * Contoh: batteryLevel 8 → VERY_LOW (karena < 10, tapi belum < 5).
+     *         batteryLevel 3 → CRITICAL
+     *         batteryLevel 25 → empty (tidak perlu notifikasi)
      */
-    public static Optional<BatteryNotificationThreshold> resolve(float batteryLevel) {
+    public static Optional<BatteryNotificationThreshold> resolve(int batteryLevel) {
         return Arrays.stream(values())
                 .sorted(Comparator.comparing(BatteryNotificationThreshold::getPercent))
                 .filter(t -> batteryLevel < t.percent)
@@ -35,7 +35,7 @@ public enum BatteryNotificationThreshold {
     /**
      * Cek apakah battery level sudah di atas semua threshold (sudah di-charge).
      */
-    public static boolean isAboveAllThresholds(float batteryLevel) {
+    public static boolean isAboveAllThresholds(int batteryLevel) {
         return Arrays.stream(values())
                 .noneMatch(t -> batteryLevel < t.percent);
     }
